+++ server/src/main/java/org/guohai/javasqlweb/beans/ServerDatabaseSyncFailure.java
@@ -1,0 +1,10 @@
+package org.guohai.javasqlweb.beans;
+
+import lombok.Data;
+
+@Data
+public class ServerDatabaseSyncFailure {
+    private Integer serverCode;
+    private String serverName;
+    private String message;
+}